Use exact segment entry lengths (including metadata) when keeping statistics. Not doing so can cause bad things to happen, especially when small objects are used and the relative error is pretty high.

+++ src/SegmentManager.cc
@@ -716,7 +716,7 @@ SegmentManager::alloc(SegletAllocator::AllocationType type,
     freeUnreferencedSegments();
 
     uint32_t slot = allocSlot(type);
-    if (slot == -1U)
+    if (slot == static_cast<uint32_t>(-1))
         return NULL;
 
     uint32_t segletSize = allocator.getSegletSize();
